% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{binomial_quantile}
\alias{binomial_quantile}
\title{Quantile of bayesian estimator for binomial trial with flat prior}
\usage{
binomial_quantile(n_positive, n_trials, p)
}
\arguments{
\item{n_positive}{number of successes in binomial trial}

\item{n_trials}{number of binomial trials}

\item{p}{probability quantile to return}
}
\description{
Using the above derivation, the bayesian estimator for the success probability p
from a binomial trial with n successes and m failures and flat prior is Beta(n+1, m+1).
}
