% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle3_support.R
\name{compute_qc_covariates}
\alias{compute_qc_covariates}
\title{Compute important quality control covariates suggested by (Lueken and Theis 2019)}
\usage{
compute_qc_covariates(cds)
}
\arguments{
\item{cds}{cell_data_set}
}
\description{
Add count_depth, n_genes, mt_fraction values to
the column data for the input cell_data_set
}
