% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_brms_flat_score_by_dose.R
\name{fit_brms_flat_score_by_dose}
\alias{fit_brms_flat_score_by_dose}
\title{brms flat model for score by compound dose}
\usage{
fit_brms_flat_score_by_dose(well_scores)
}
\arguments{
\item{well_scores}{tibble::tibble as output by read_well_scores}
}
\value{
list of brms::brmsfit objects one for each compound
}
\description{
Use the R BRMS package to fit dose response data to with constant function
}
