% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_MuSyC_score_by_dose.R
\name{fit_MuSyC_score_by_dose_robust}
\alias{fit_MuSyC_score_by_dose_robust}
\title{Fit the MuSyC synergy model by dose}
\usage{
fit_MuSyC_score_by_dose_robust(
  well_scores,
  group_vars = vars(compound),
  prior = MuSyC_default_prior(),
  inits = MuSyC_default_inits,
  combine = FALSE,
  verbose = FALSE,
  iter = 8000,
  cores = 4,
  stan_model_args = list(verbose = FALSE),
  control = list(adapt_delta = 0.99, max_treedepth = 12),
  model_evaluation_criteria = c("loo", "bayes_R2"),
  ...
)
}
\arguments{
\item{well_scores}{data.frame
with columns: dose1, dose2, d1_scale_factor, d2_scale_factor, n_positive, count, [<group_vars>]}

\item{group_vars}{quosures list
dplyr::vars(...) columns to over when fitting synergy model}

\item{combine}{combine the grouped models into a single brms model}

\item{verbose}{verbose output}

\item{iter}{number of stan NUTS sampling steps}

\item{cores}{number of cores used for sampling}

\item{stan_model_args}{stan model arguments}

\item{control}{stan control arguments

The

   bernoulli_inf(n_positive / count) = Ed ~ MuSyC(d1, d2, C_params, E_params, s_params, alpha)

   To improve numeric stability, the d1 and d2 and C1 and C2 variables
   are scaled to improve numeric stability:

      d1 = dose1/max(dose1)
      d2 = dose2/max(dose2)
      drug1_IC50 = C1 * max(dose1)
      drug2_IC50 = C2 * max(dose2)

Functional form:
Ed ~ 
Ed ~ (
       C1^h1 * C2^h2 * E0 +
       d1^h1 * C2^h2 * E1 +
       C1^h1 * d2^h2 * E2 +
       d1^h1 * d2^h2 * E3 * alpha
     ) / (
       C1^h1 * C2^h2 +
       d1^h1 * C2^h2 +
       C1^h1 * d2^h2 +
       d1^h1 * d2^h2 * alpha
     )

The ratio in the functional form is numerically unstable, transform to use the log_sum_exp trick,
Let

   numerator_parts = [
     h1*log(C1) + h2*log(C2) + log(E0),
     h1*log(d1) + h2*log(C2) + log(E1),
     h1*log(C1) + h2*log(d2) + log(E2),
     h1*log(d1) + h2*log(d2) + log(E3) + log(alpha)]
 
and    
   
   denominator_parts = [
     h1*log(C1) + h2*log(C2),
     h1*log(d1) + h2*log(C2),
     h1*log(C1) + h2*log(d2),
     h1*log(d1) + h2*log(d2)]     

Then 

   Ed ~ exp(log_sum_exp(numerator_parts) - log_sum_exp(denominator_parts)).




##############################################
# Proof of the definitions of the parameters #
##############################################

Claim: When d1=0 and d2=0 then Ed = E0
Ed = (C1^h1 * C2^h2 * E0) / (C1^h1 * C2^h2)
   = E0

Claim: When d1=0 and d2 -> Inf then Ed = E2
Ed = (C2^h2 * E0 + d2^h2 * E2) / (C2^h2 + d2^h2)
   = (d2^h2 * E2) / (d2^h2)
   = E2

When d1>0 and d2 -> Inf then Ed
Ed = (C1^h1 * C2^h2 * E0 + C1^h1 * C2^h2 * E2) / (C1^h1 * C2^h2 + C1^h1 * C2^h2)
   = (E0 + E2) / 2}

\item{C1_prior}{prior distribution for Ed when d1=d1_IC50, d2=0}

\item{C2_prior}{prior distribution for Ed when d1=0, d2=d2_IC50}

\item{s1_prior}{prior distribution for d(Ed)/d(d1) when d1=d1_IC50, d2=0}

\item{s2_prior}{prior distribution for d(Ed)/d(d2) when d1=0, d2=d2_IC50}

\item{log10alpha_prior}{prior distribution for alpha synergy parameter}

\item{E0_prior}{prior distribution for Ed when d1=0, d2=0}

\item{E1_prior}{prior distribution for Ed when d1=Inf, d2=0}

\item{E2_prior}{prior distribution for Ed when d1=0, d2=Inf}

\item{E3_alpha_prior}{prior distribution for Ed scaled by alpha when d1=Inf, d2=Inf}

\item{C1_init}{initial sampling distribution for the C1 parameter}

\item{C2_init}{initial sampling distribution for the C2 parameter}

\item{s1_init}{initial sampling distribution for the s1 parameter}

\item{s2_init}{initial sampling distribution for the s2 parameter}

\item{log10alpha_init}{intial sampling distribution for the alpha parameter}

\item{E0_init}{initial sampling distribution for the E0 parameter}

\item{E1_init}{initial sampling distribution for the E1 parameter}

\item{E2_init}{initial sampling distribution for the E2 parameter}

\item{E3_alpha_init}{initial sampling distribution for the E3 parameter}
}
\description{
Fit the MuSyC model with a binomial response
   The
}
