% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_brms_hill_score_by_dose.R
\name{generate_hill_effects}
\alias{generate_hill_effects}
\title{Simulate hill model from the parameters}
\usage{
generate_hill_effects(log_dose, top, bottom, ic50, hill)
}
\arguments{
\item{log_dose}{log10(doses) at which to evaluate the model}

\item{top}{top parameter}

\item{bottom}{bottom parameter}

\item{ic50}{ic50 parameter}

\item{hill}{hill parameter}
}
\value{
vector of values
}
\description{
Simulate hill model from the parameters
}
