% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_indicator.R
\docType{data}
\name{GeomIndicator}
\alias{GeomIndicator}
\title{ggplot2 geom that adds a label to facet panes}
\format{
An object of class \code{GeomIndicator} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomIndicator
}
\description{
ggplot2 geom that adds a label to facet panes
}
\keyword{datasets}
