% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_support.R
\name{get_database_connection}
\alias{get_database_connection}
\title{Construct a database from database connection parameters}
\usage{
get_database_connection(database_parameters)
}
\arguments{
\item{database_parameters}{List with the following elements
database_type, host, port, user, password, schema}
}
\value{
DBI database connect with the given schema specified
}
\description{
It's recommended to store database connection parameters in the
   <project>/parameters.R file
}
