% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_workflow.R
\name{loo_R2_indicator}
\alias{loo_R2_indicator}
\title{Summarize the loo_R2 as a plot indicator}
\usage{
loo_R2_indicator(model, digits = 2, ...)
}
\arguments{
\item{model}{brmsfit}

\item{digits}{number of significant figure digits}
}
\value{
MPStats::geom_indicator object to that can be added to a ggplot2::ggplot
}
\description{
Produces a label on the plot using the coordinates of the viewport rather than
  the plot coordinates. The indicator is with the indicated number of digits as sig figs.
}
\details{
R^2 = mean [<Q2.5>, <Q97.5>]
}
