% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cell_count_by_batch_vars.R
\name{model_cell_count_by_batch_vars_lm}
\alias{model_cell_count_by_batch_vars_lm}
\title{Model cell count by batch variables using a linear model}
\usage{
model_cell_count_by_batch_vars_lm(well_scores)
}
\arguments{
\item{well_scores}{result of read_well_scores()}
}
\value{
"lm" object of model
}
\description{
Here using a sqrt() transform to make the response more normal
}
