% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_score_by_batch_vars.R
\name{model_score_by_batch_vars_lm}
\alias{model_score_by_batch_vars_lm}
\title{Model score by batch variables using a weighted least squares with linear predictors}
\usage{
model_score_by_batch_vars_lm(well_scores)
}
\arguments{
\item{well_scores}{result of read_well_scores()}
}
\value{
"lm" object of model
}
\description{
Model score by batch variables using a weighted least squares with linear predictors
}
