% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_MuSyC_score_by_dose.R
\name{MuSyC_default_prior}
\alias{MuSyC_default_prior}
\title{Default priors for MuSyC synergy model}
\usage{
MuSyC_default_prior()
}
\value{
named vector of class 'brms::brmsprior'
}
\description{
Default priors for MuSyC synergy model
}
