% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_workflow.R
\name{pairsplot}
\alias{pairsplot}
\title{Customized pairs plot}
\usage{
pairsplot(model, size = 0.2, alpha = 0.1, ...)
}
\arguments{
\item{model}{brmsfit model}
}
\value{
ggplot2::ggplot object
}
\description{
Customise pairs by
  * adjusting the default off-diagonal points to be smaller and more transperent
  * adding a title that uses the model$name field
}
