% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_layout.R
\name{plate_layout}
\alias{plate_layout}
\title{get plate layout}
\usage{
plate_layout(well_scores, week, plate)
}
\arguments{
\item{well_scores}{well scores data.frame as output by read_well_scores()}

\item{week}{week identifier (1-based integer)}

\item{plate}{plate index (1-based index)}
}
\value{
Table with row and columns as on the plate
       with cell values 'compound_id|log_dose'
}
\description{
get plate layout
}
