% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_binomial_trial_credible_intervals.R
\name{plot_binomial_variance_by_cell_count}
\alias{plot_binomial_variance_by_cell_count}
\title{A diagnostic plot for error analysis for binomial trials}
\usage{
plot_binomial_variance_by_cell_count()
}
\value{
ggplot2 plot of probability positive by cell count
}
\description{
A diagnostic plot for error analysis for binomial trials
}
