% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_brms_predictor_residuals_score_by_dose.R
\name{plot_brms_predictor_residual_score_by_dose}
\alias{plot_brms_predictor_residual_score_by_dose}
\title{Plot the score as a function of compound dose for brms models}
\usage{
plot_brms_predictor_residual_score_by_dose(
  well_scores,
  flat_model,
  hill_model,
  subtitle = NULL
)
}
\arguments{
\item{well_scores}{tibble::tibble as output by read_well_scores}

\item{flat_model}{tibble::tibble as output by fit_brms_flat_score_by_dose}

\item{hill_model}{tibble::tibble as output by fit_brms_hill_score_by_dose}

\item{subtitle}{string subtitle for plot}
}
\value{
ggplot2 plot
}
\description{
Plot the score as a function of compound dose for brms models
}
