% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_brms_score_by_dose.R
\name{plot_brms_score_by_dose}
\alias{plot_brms_score_by_dose}
\title{Plot the score as a function of compound dose for brms models}
\usage{
plot_brms_score_by_dose(well_scores, flat_model, hill_model,
  subtitle = NULL)
}
\arguments{
\item{well_scores}{tibble::tibble as output by read_well_scores}

\item{subtitle}{string subtitle for plot}

\item{fits}{tibble::tibble as output by fit_drc_score_by_dose}
}
\value{
ggplot2 plot
}
\description{
For each compound plot a score by log_dose facet panel with
 *  the scores for each dose shown as the pooled score across replicates
 *  error bars as the standard deviation of the per-dose Bayesian uncertainty with a flat prior
 *  plot the following models
 *     flat model
 #     4-parameter sigmoid with bottom fixed to 0.
 *  labeled with the p-value of a constant-value fit
}
