% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_checkerboard_score_by_dose.R
\name{plot_checkerboard_score_by_dose}
\alias{plot_checkerboard_score_by_dose}
\title{Plot dose-response checkerboard}
\usage{
plot_checkerboard_score_by_dose(well_scores, contour_color = "gold")
}
\arguments{
\item{well_scores}{data.frame with columns [dose_1, dose_2, score]}

\item{contour_color}{}
}
\description{
Plot dose-response checkerboard
}
