% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_embedded_cluster_principal_curves.R
\name{plot_embedded_cluster_principal_curves}
\alias{plot_embedded_cluster_principal_curves}
\title{Plot embedded cluster principal cluves}
\usage{
cluster_principal_curves <- dplyr::bind_cols(
   arrow::read_parquet("input/umap_embedding.parquet"),
   arrow::read_parquet("input/hbscan_clustering.parquet")) %>%
   MPStats::fit_principal_curves() %>%
   dplyr::arrange(princurve_order)

plot <- MPStats::plot_embedded_principal_curves(cluster_principal_curves)
ggplot2::ggsave(
  plot=plot,
  filename=paste0("product/embedded_principal_curves_", MPStats::date_code(), ".pdf"),
  width=6, height=6,
  useDingbats=FALSE)
ggplot2::ggsave(
  plot=dye_plot,
  filename=paste0("product/embedded_principal_curves_", MPStats::date_code(), ".png"),
  width=6, height=6)
}
\arguments{
\item{subtitle}{plot subtitle}

\item{cluster_curves}{tibble::tibble with columns
UMAP_1  :         coordinate 1 of the umap embedding
UMAP_2  :         coordiante 2 of the umap embedding
cluster_id :      cluster identifier
princurve_1 :     coordinate 1 of the principal curve projection
princurve_2 :     coordiante 2 of the principal curve projection
princurve_order : order of points along the principal curve projection}
}
\description{
This makes a UMAP density plot and adds principal curves to each cluster
}
