% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_correlation_matrix.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot feature heatmap}
\usage{
For example, to plot the correlation


  heatmap_args <- cell_features %>%
    stats::cor() %>%
    MPStats::plot_heatmap()
  pdf(fname, height=height, width=width)
  suppressWarnings(ret <- do.call(gplots::heatmap.2, heatmap_args))
}
\arguments{
\item{x}{matrix to be plotted}
}
\value{
gplots object with a matrix where cell (i,j) is colored by the in the cell
}
\description{
plot feature heatmap
}
