% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score_by_batch_vars.R
\name{plot_score_by_batch_vars}
\alias{plot_score_by_batch_vars}
\title{plot cell count by batch variables}
\usage{
plot_score_by_batch_vars(well_scores, subtitle = NULL)
}
\arguments{
\item{well_scores}{tibble::tibble as output by read_well_scores}

\item{subtitle}{plot subtitle, typically the study identifier}
}
\value{
ggplot2 object with the lattice plot with a panel for each batch variable
}
\description{
plot cell count by batch variables
}
