% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score_by_batch_vars_scatter.R
\name{plot_score_by_batch_vars_scatter}
\alias{plot_score_by_batch_vars_scatter}
\title{plot cell count by batch variables as a scatter plot}
\usage{
plot_score_by_batch_vars_scatter(well_scores, subtitle = NULL)
}
\arguments{
\item{well_scores}{tibble::tibble as output by read_well_scores}

\item{subtitle}{plot subtitle, typically the study identifier}
}
\value{
ggplot2 object with the lattice plot with a panel for each batch variable
}
\description{
plot cell count by batch variables as a scatter plot
}
