% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score_by_cell_count.R
\name{plot_score_by_cell_count}
\alias{plot_score_by_cell_count}
\title{plot score by cell count}
\usage{
plot_score_by_cell_count(well_scores, subtitle = NULL)
}
\arguments{
\item{well_scores}{tibble::tibble as output by read_well_scores}

\item{subtitle}{plot subtitle, typically the study identifier}
}
\value{
ggplot2 object
}
\description{
plot score by cell count
}
