% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Zprime_by_top_features.R,
%   R/plot_top_features_by_controls.R
\name{plot_Zprime_by_top_features}
\alias{plot_Zprime_by_top_features}
\title{Plot Zprime by top features}
\usage{
plot_Zprime_by_top_features(cell_features, top_k_features = 4)

plot_Zprime_by_top_features(cell_features, top_k_features = 4)
}
\description{
For each object, plot the top 5 feature by Zprime colored by the
dye used to identify it

For each object, plot the top 4 feature by Zprime colored by the
dye used to identify it
}
