% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle3_support.R
\name{populate_cds}
\alias{populate_cds}
\title{Popluate a Monocle3 cell data set from cell features}
\usage{
populate_cds(
  cell_features,
  cell_feature_columns,
  cell_metadata_columns,
  embedding_type = c("UMAP"),
  embedding = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cell_features}{data.frame:
rows: cells, columns: features}

\item{cell_metadata_columns}{data.frame:
rows: features, columns feature metadata
Note that there must be a column named `feature`}

\item{cell_metadata}{rows: cells, columns cell metadata}
}
\description{
The Monocle3 cell data set is a container
}
