% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_workflow.R
\name{prior_posterior_plot}
\alias{prior_posterior_plot}
\title{Create a plot that compares the distribution of prior and posterior
marginal distributions for each variable}
\usage{
prior_posterior_plot(model)
}
\arguments{
\item{model}{brmsfit model}
}
\value{
ggplot2::ggplot object
}
\description{
To make the plot managable, the distribution for each variable
is trimmed at the 95% IQR
}
\details{
The plot title uses `model$name`
}
