% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_workflow.R
\name{rankplot}
\alias{rankplot}
\title{Customized rank plot}
\usage{
rankplot(model)
}
\arguments{
\item{model}{brmsfit model}
}
\value{
ggplot2::ggplot object
}
\description{
Customise bayesplot::mcmc_rank_overlay by
  * adding a title that uses the model$name field
}
