% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cell_features.R
\name{read_cell_features}
\alias{read_cell_features}
\title{Read cell features}
\usage{
read_cell_features(input)
}
\arguments{
\item{input}{path to cell features csv (e.g. AGM_FeatureSet_RAW55.csv) or}
}
\value{
cell_features tibble::tibble with columns
    COND
    Texture*
    RadialDistribution*
    Intensity*
    Granularity*
    Correlation*
    AreaShape*
}
\description{
Read cell features
}
