% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_compound_moa.R
\name{read_compound_moa}
\alias{read_compound_moa}
\title{Read well scores table}
\usage{
read_compound_moa(file)
}
\arguments{
\item{filename}{path to compound mechanism of action csv, (e.g. AGM_moa.csv)}
}
\value{
compound_moa tibble::tibble with columns
  Classes 'spec_tbl_df', 'tbl_df', 'tbl' and 'data.frame':	40 obs. of  3 variables:
    $ compound      : chr  "PP-2" "emetine" "AZ258" "cytochalasin B" ...
    $ concentration : num  3 0.3 1 10 3 0.01 0.3 3 0.3 0.3 ...
    $ moa           : chr  "Epithelial" "Protein synthesis" "Aurora kinase inhibitors" "Actin disruptors" ...
    $ moa2          : chr  "other" "other" "other" "cytoskeleton" ...
}
\description{
Read well scores table
}
