% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Zprime_by_plate.R
\name{read_Zprime_by_plate}
\alias{read_Zprime_by_plate}
\title{Read Zprime by plate}
\usage{
read_Zprime_by_plate(input)
}
\arguments{
\item{input}{path to Zprime csv (e.g. Master55.csv) or
data.frame with columns
    Metadata_PlateID_Nuclei               : chr
    Metadata_WellID_Nuclei                : chr
    Children_Cells_Count_Nuclei           : num
    Metadata_Image_Metadata_Compound      : chr
    Metadata_Image_Metadata_Concentration : num
    probPos                               : num
    probNeg                               : num}
}
\value{
well_scores data.frame with columns
  week_id    : int  10 10 10 10 10 10 10 10 10 10 ...
  plate_ind  : int  1 1 1 1 1 1 1 1 1 1 ...
  row        : int  2 2 2 2 2 2 2 2 2 2 ...
  column     : int  2 3 4 5 6 7 8 9 10 11 ...
  compound   : chr  "DMSO" "Cdk1/2 inhibitor (NU6102)" "Cdk1/2 inhibitor (NU6102)" ...
  log_dose   : num  -Inf 4 3.48 3 2.48 ...
  is_control : logi  TRUE FALSE FALSE FALSE FALSE FALSE ...
  cell_count : num  1442 479 743 888 951 ...
  n_positive : int  0 0 0 0 0 0 0 0 0 303 ...
}
\description{
Read Zprime by plate
}
