% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle3_support.R
\name{serialize_clusters}
\alias{serialize_clusters}
\title{Write out clusters from monocle3 cell dataset}
\usage{
serialize_clusters(
  cds,
  output_fname,
  reduction_method = c("UMAP", "tSNE", "PCA", "LSI", "Aligned"),
  verbose = FALSE
)
}
\arguments{
\item{cds}{monocle3 cell dataset}

\item{output_fname}{path to output parquet file where the clusters should be written
the data has a single column [cluster_label] with values for each object}

\item{reduction_method}{the reduction method for which the clusters were computed [default: UMAP]}

\item{verbose}{verbose output [default: FALSE]}
}
\description{
Write out clusters from monocle3 cell dataset
}
