% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_workflow.R
\name{traceplot}
\alias{traceplot}
\title{Customized trace plot}
\usage{
traceplot(model, ...)
}
\arguments{
\item{model}{brmsfit model}
}
\value{
ggplot2::ggplot object
}
\description{
Customise bayesplot::mcmc_traceplot by
  * adding a red smoothing line for each trace.
  * adding a title that uses the model$name field
}
