% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{umich_colors}
\alias{umich_colors}
\title{Univrsity of Michigan colors}
\format{An object of class \code{character} of length 18.}
\usage{
umich_colors
}
\description{
https://med.umich.edu/branding/color.html
}
\keyword{datasets}
