% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{well_id_to_column}
\alias{well_id_to_column}
\title{Extract column index from well identifier}
\usage{
well_id_to_column(well_id)
}
\arguments{
\item{well_id}{well index of the form e.g. 'H14'}
}
\value{
integer index corresponding to the column of the well
}
\description{
Convert e.g. 'H14' well identifier to 14 because '14' is the 14th row.
}
