% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{well_id_to_row}
\alias{well_id_to_row}
\title{Extract row index from well identifier}
\usage{
well_id_to_row(well_id)
}
\arguments{
\item{well_id}{well index of the form e.g. 'H14'}
}
\value{
integer index corresponding to the row of the well
}
\description{
Convert e.g. 'H14' well identifier to 8 because 'H' is the 8th row.
}
