% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{well_index_to_row}
\alias{well_index_to_row}
\title{Extract row index from well index}
\usage{
well_index_to_row(well_index)
}
\arguments{
\item{well_index}{well index of the form e.g. '224'}
}
\value{
integer index corresponding to the row of the well
}
\description{
Convert e.g. '224' well index to 10 because it is on the 10th row in a 384 well plate
}
