% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Zprime_by_top_features.R
\name{Zprime}
\alias{Zprime}
\title{Compute Zprime score}
\usage{
Zprime(positives, negatives)
}
\description{
J. H. Zhang, T. D. Chung, K. R. Oldenburg. A Simple Statistical Parameter for Use in Evaluation and Validation of High Throughput Screening Assays. J Biomol Screening, 1999.
}
