% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cell_count_by_batch_vars_density.R
\name{plot_cell_count_by_batch_vars_density}
\alias{plot_cell_count_by_batch_vars_density}
\title{plot cell count by batch variables as a density plot}
\usage{
plot_cell_count_by_batch_vars_density(well_scores, subtitle = NULL)
}
\arguments{
\item{well_scores}{tibble::tibble as output by read_well_scores}

\item{subtitle}{plot subtitle, typically the study identifier}
}
\value{
ggplot2 object with the lattice plot with a panel for each batch variable
}
\description{
For each batch variable, show a facet panel with sqrt(cell_count) by batch variable
}
