% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Zprime_by_dye.R
\name{plot_Zprime_by_dye}
\alias{plot_Zprime_by_dye}
\title{Plot Zprime score by Dye combination}
\usage{
plot_Zprime_by_dye(Zprime_by_plate, subtitle = NULL)
}
\arguments{
\item{Zprime_by_plate}{tibble::tibble}

\item{subtitle}{string subtitle for plot}
}
\value{
ggplot2 plot
}
\description{
Plot Zprime score by Dye combination
}
