% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_compound_moa.R
\name{read_compound_moa}
\alias{read_compound_moa}
\title{Read compound mechanism of action table}
\usage{
read_compound_moa(input)
}
\arguments{
\item{input}{path to compound mechanism of action csv, (e.g. AGM_moa.csv) or
data.frame with columns
    compound      : chr    # name of compound
    concentration : num    # 
    moa           : chr    # detailed mechanism of action
    moa2          : chr    # class of mechanism of action}
}
\value{
compound_moa tibble::tibble with columns
  Classes 'spec_tbl_df', 'tbl_df', 'tbl' and 'data.frame':	40 obs. of  3 variables:
    $ compound      : chr  "PP-2" "emetine" "AZ258" "cytochalasin B" ...
    $ concentration : num  3 0.3 1 10 3 0.01 0.3 3 0.3 0.3 ...
    $ moa           : chr  "Epithelial" "Protein synthesis" "Aurora kinase inhibitors" "Actin disruptors" ...
    $ moa2          : chr  "other" "other" "other" "cytoskeleton" ...
}
\description{
Read compound mechanism of action table
}
